+++ ChangeLog.md
@@ -1,6 +1,15 @@
 # Changelog for wallet-proxy
 
 ## Unreleased changes
+
+## 0.13.4
+ - add `GET /v0/bakerPool/{bakerId}` endpoint
+ - add `GET /v0/chainParameters` endpoint
+ - add `GET /v0/nextPayday` endpoint
+ - add support for delegation events and new baker and delegation transactions
+   in the `submissionStatus` endpoint.
+ - add support for the cost of the new baker and delegation transactions
+   in the `transactionCost` endpoint.
  - add support for events triggered by V1 smart contracts
 
 ## 0.8.2